feat: chunked upload with progress, retry, and resume for large files
- Chunked upload API: /api/upload/init, /api/upload/chunk,
  /api/upload/status/{id}, /api/upload/complete
- Files >=5MB are split into 5MB chunks, each uploaded separately
- Auto-retry per chunk with exponential backoff (3 attempts)
- Resume: upload ID saved to localStorage, on page refresh the
  client queries server for received chunks and only uploads
  missing ones
- Progress bar UI with percentage and bytes transferred
- "Resume" button appears if all retries fail for a chunk
- Small files (<5MB) use simple XHR upload with progress
- Max upload size configurable via MAX_UPLOAD_MB env var (default 2GB)
- Chunk sweeper cleans up incomplete uploads after 1 hour
- REST API docs updated with chunked upload endpoints
- Dockerfile includes MAX_UPLOAD_MB=2048 default

+++ mcp.go
@@ -355,7 +355,7 @@ func mcpUploadFile(filename, content, mimeType, ttlStr string) (interface{}, *MC
 	if err != nil {
 		return nil, &MCPError{Code: -32602, Message: "invalid base64 content"}
 	}
-	if len(data) > maxUploadBytes {
+	if int64(len(data)) > maxUploadBytes {
 		return nil, &MCPError{Code: -32602, Message: fmt.Sprintf("file too large (max %d MB)", maxUploadMB)}
 	}
